Refactor navigation and profile components; add My Data link and enhance profile picture upload functionality
