+++ chd-web/src/pages/my-data.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
 import Head from 'next/head';
@@ -6,10 +6,19 @@ import { useAuth } from '../lib/supabaseClient';
 import Nav from '../components/Nav';
 import Footer from '../components/Footer';
 import Button from '../components/ui/Button';
+import { listUserCSVs, downloadCSV, type UserCsvObject } from '../lib/storage';
+import Papa from 'papaparse';
 
 export default function MyData() {
   const router = useRouter();
   const { user, loading } = useAuth();
+  const [rows, setRows] = useState<UserCsvObject[] | null>(null);
+  const [err, setErr] = useState<string | null>(null);
+  const [downloading, setDownloading] = useState<string | null>(null);
+  const [combined, setCombined] = useState<Record<string, string>[] | null>(null);
+  const [isCombining, setIsCombining] = useState(false);
+  const [page, setPage] = useState(1);
+  const pageSize = 25;
 
   useEffect(() => {
     // Redirect to login if not authenticated
@@ -17,6 +26,19 @@ export default function MyData() {
       router.push('/login?next=/my-data');
     }
   }, [user, loading, router]);
+
+  useEffect(() => {
+    const load = async () => {
+      if (!user) return;
+      try {
+        const list = await listUserCSVs(user);
+        setRows(list);
+      } catch (e: any) {
+        setErr(e?.message || 'Failed to load your files');
+      }
+    };
+    load();
+  }, [user]);
 
   // Show loading state
   if (loading) {
@@ -49,123 +71,158 @@ export default function MyData() {
 
         <main className="flex-grow bg-gradient-to-br from-red-50 to-gray-50 py-16">
           <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            {/* Page Header */}
-            <div className="text-center mb-12">
-              <h1 className="text-4xl font-bold text-gray-900 mb-4">
-                My Data
-              </h1>
-              <p className="text-xl text-gray-600">
-                Your saved analyses and predictions
-              </p>
+            <div className="mb-8 text-center">
+              <h1 className="text-4xl font-bold text-gray-900 mb-2">My Data</h1>
+              <p className="text-gray-600">Your saved analyses and predictions</p>
             </div>
 
-            {/* Main Content Card */}
-            <div className="max-w-3xl mx-auto">
-              <div className="bg-white rounded-2xl shadow-xl border border-gray-100 p-12">
-                {/* Icon */}
-                <div className="text-center mb-8">
-                  <div className="inline-flex items-center justify-center w-20 h-20 rounded-full bg-red-100 mb-6">
-                    <svg 
-                      className="w-10 h-10 text-red-600" 
-                      fill="none" 
-                      viewBox="0 0 24 24" 
-                      stroke="currentColor"
-                    >
-                      <path 
-                        strokeLinecap="round" 
-                        strokeLinejoin="round" 
-                        strokeWidth={2} 
-                        d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" 
-                      />
-                    </svg>
-                  </div>
-                  
-                  <h2 className="text-2xl font-bold text-gray-900 mb-3">
-                    Coming Soon
-                  </h2>
-                  
-                  <p className="text-lg text-gray-600 mb-2">
-                    Your saved results will appear here
-                  </p>
-                  
-                  <p className="text-gray-500 max-w-md mx-auto">
-                    We&apos;re working on bringing you the ability to save, view, and manage 
-                    your cardiovascular risk assessments. Stay tuned!
-                  </p>
+            {/* Saved CSVs */}
+            <div className="bg-white rounded-2xl shadow-lg border border-gray-100 p-6 mb-8">
+              <div className="flex items-center justify-between mb-4">
+                <h2 className="text-xl font-semibold text-gray-900">Saved CSV Files</h2>
+                <Link href="/predictor"><Button variant="secondary" size="sm">New Assessment</Button></Link>
+              </div>
+              {err && (
+                <div className="mb-4 p-3 bg-red-50 border border-red-200 rounded-lg text-red-700">{err}</div>
+              )}
+              {rows && rows.length > 0 ? (
+                <div className="overflow-x-auto">
+                  <table className="min-w-full text-sm">
+                    <thead>
+                      <tr className="text-left text-gray-600">
+                        <th className="px-3 py-2 border-b">File</th>
+                        <th className="px-3 py-2 border-b">Updated</th>
+                        <th className="px-3 py-2 border-b">Size</th>
+                        <th className="px-3 py-2 border-b text-right">Actions</th>
+                      </tr>
+                    </thead>
+                    <tbody>
+                      {rows.map((r) => (
+                        <tr key={r.path}>
+                          <td className="px-3 py-2 border-b">{r.name}</td>
+                          <td className="px-3 py-2 border-b">{r.updated_at ? new Date(r.updated_at).toLocaleString() : '—'}</td>
+                          <td className="px-3 py-2 border-b">{typeof r.size === 'number' ? `${(r.size/1024).toFixed(1)} KB` : '—'}</td>
+                          <td className="px-3 py-2 border-b text-right">
+                            <Button
+                              variant="primary"
+                              size="sm"
+                              isLoading={downloading === r.path}
+                              onClick={async () => {
+                                try {
+                                  setDownloading(r.path);
+                                  const text = await downloadCSV(r.path);
+                                  const blob = new Blob([text], { type: 'text/csv;charset=utf-8' });
+                                  const url = URL.createObjectURL(blob);
+                                  const a = document.createElement('a');
+                                  a.href = url;
+                                  a.download = r.name || 'data.csv';
+                                  document.body.appendChild(a);
+                                  a.click();
+                                  document.body.removeChild(a);
+                                  URL.revokeObjectURL(url);
+                                } finally {
+                                  setDownloading(null);
+                                }
+                              }}
+                            >Download</Button>
+                          </td>
+                        </tr>
+                      ))}
+                    </tbody>
+                  </table>
                 </div>
+              ) : (
+                <div className="text-gray-600">No files yet. Run a single or batch prediction to save results.</div>
+              )}
+            </div>
 
-                {/* User Info */}
-                <div className="bg-gray-50 rounded-xl p-6 mb-8">
-                  <div className="flex items-center">
-                    <div className="flex-shrink-0">
-                      <div className="w-12 h-12 rounded-full bg-red-600 flex items-center justify-center">
-                        <span className="text-white font-semibold text-lg">
-                          {user.email?.[0].toUpperCase()}
-                        </span>
+            {/* Combined Table */}
+            {rows && rows.length > 0 && (
+              <div className="bg-white rounded-2xl shadow-lg border border-gray-100 p-6">
+                <div className="flex items-center justify-between mb-4">
+                  <h2 className="text-xl font-semibold text-gray-900">All Data (Combined)</h2>
+                  <div className="flex items-center gap-2">
+                    <Button
+                      variant="primary"
+                      size="sm"
+                      isLoading={isCombining}
+                      onClick={async () => {
+                        if (!rows) return;
+                        setIsCombining(true);
+                        setErr(null);
+                        try {
+                          const all: Record<string, string>[] = [];
+                          for (const r of rows) {
+                            const text = await downloadCSV(r.path);
+                            const parsed = Papa.parse<Record<string, string>>(text, { header: true, skipEmptyLines: true });
+                            if (parsed.data) all.push(...parsed.data);
+                          }
+                          setCombined(all);
+                          setPage(1);
+                        } catch (e: any) {
+                          setErr(e?.message || 'Failed to build combined table');
+                        } finally {
+                          setIsCombining(false);
+                        }
+                      }}
+                    >Load Combined Table</Button>
+                    {combined && combined.length > 0 && (
+                      <Button
+                        variant="secondary"
+                        size="sm"
+                        onClick={() => {
+                          const csv = Papa.unparse(combined);
+                          const blob = new Blob([csv], { type: 'text/csv;charset=utf-8' });
+                          const url = URL.createObjectURL(blob);
+                          const a = document.createElement('a');
+                          a.href = url;
+                          a.download = 'heartsense_combined.csv';
+                          document.body.appendChild(a);
+                          a.click();
+                          document.body.removeChild(a);
+                          URL.revokeObjectURL(url);
+                        }}
+                      >Export CSV</Button>
+                    )}
+                  </div>
+                </div>
+                {combined && (
+                  combined.length > 0 ? (
+                    <div className="overflow-x-auto">
+                      <table className="min-w-full text-sm">
+                        <thead>
+                          <tr className="text-left text-gray-600">
+                            {Object.keys(combined[0]).slice(0, 30).map((h) => (
+                              <th key={h} className="px-3 py-2 border-b">{h}</th>
+                            ))}
+                          </tr>
+                        </thead>
+                        <tbody>
+                          {combined.slice((page-1)*pageSize, page*pageSize).map((row, idx) => (
+                            <tr key={idx}>
+                              {Object.keys(combined[0]).slice(0, 30).map((h) => (
+                                <td key={h} className="px-3 py-2 border-b">{row[h] ?? ''}</td>
+                              ))}
+                            </tr>
+                          ))}
+                        </tbody>
+                      </table>
+                      <div className="flex items-center justify-between mt-4 text-sm text-gray-600">
+                        <div>
+                          Showing {(page-1)*pageSize+1}-{Math.min(page*pageSize, combined.length)} of {combined.length}
+                        </div>
+                        <div className="flex items-center gap-2">
+                          <Button variant="ghost" size="sm" disabled={page===1} onClick={() => setPage((p)=>Math.max(1,p-1))}>Prev</Button>
+                          <Button variant="ghost" size="sm" disabled={page*pageSize>=combined.length} onClick={() => setPage((p)=>p+1)}>Next</Button>
+                        </div>
                       </div>
                     </div>
-                    <div className="ml-4">
-                      <p className="text-sm text-gray-500">Signed in as</p>
-                      <p className="text-base font-medium text-gray-900">{user.email}</p>
-                    </div>
-                  </div>
-                </div>
-
-                {/* Feature Preview */}
-                <div className="space-y-4 mb-8">
-                  <h3 className="text-lg font-semibold text-gray-900 mb-4">
-                    Planned Features:
-                  </h3>
-                  
-                  <div className="flex items-start">
-                    <svg className="w-5 h-5 text-green-500 mt-0.5 mr-3" fill="currentColor" viewBox="0 0 20 20">
-                      <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                    </svg>
-                    <p className="text-gray-700">Save and track your risk assessments over time</p>
-                  </div>
-                  
-                  <div className="flex items-start">
-                    <svg className="w-5 h-5 text-green-500 mt-0.5 mr-3" fill="currentColor" viewBox="0 0 20 20">
-                      <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                    </svg>
-                    <p className="text-gray-700">Export your results as PDF or CSV</p>
-                  </div>
-                  
-                  <div className="flex items-start">
-                    <svg className="w-5 h-5 text-green-500 mt-0.5 mr-3" fill="currentColor" viewBox="0 0 20 20">
-                      <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                    </svg>
-                    <p className="text-gray-700">View trends and changes in your health metrics</p>
-                  </div>
-                  
-                  <div className="flex items-start">
-                    <svg className="w-5 h-5 text-green-500 mt-0.5 mr-3" fill="currentColor" viewBox="0 0 20 20">
-                      <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                    </svg>
-                    <p className="text-gray-700">Share reports securely with healthcare providers</p>
-                  </div>
-                </div>
-
-                {/* CTA Buttons */}
-                <div className="flex flex-col sm:flex-row gap-4">
-                  <Link href="/predictor" className="flex-1">
-                    <Button variant="primary" size="lg" className="w-full">
-                      Run New Assessment
-                    </Button>
-                  </Link>
-                  
-                  <Button 
-                    variant="secondary" 
-                    size="lg" 
-                    className="flex-1"
-                    disabled
-                    title="Export feature coming soon"
-                  >
-                    Export Data
-                  </Button>
-                </div>
+                  ) : (
+                    <div className="text-gray-600">No rows found in your CSVs.</div>
+                  )
+                )}
               </div>
-            </div>
+            )}
           </div>
         </main>
